+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Implement `Compony::Components::Show`
+
 # 0.5.9
 
 - Fix a bug in generator that was introduced with the previous commit